Brand

+++ src/main/java/com/rent/entities/concretes/Brand.java
@@ -1,13 +1,13 @@
 package com.rent.entities.concretes;
 
 public class Brand {
-    private Long id;
+    private int id;
     private String name;
 
     public Brand() {
     }
 
-    public Brand(Long id, String name) {
+    public Brand(int id, String name) {
         this.id = id;
         this.name = name;
     }
